Player 2 is using Decision Matrix strategy, want to test new numbers and graph more game sets

+++ Decision_Matrix.py
@@ -27,15 +27,17 @@ for x in range(n):
     for y in range(n):
         for z in range(n):
             for a in range(n):
-                decision_array[x,y,z,a] = math.ceil(x/2)
+                decision_array[x,y,z,a] = int(math.ceil(x/2))
                 
 #defining decision function
 def decision(self_creatures=0, opponent_creatures=0, self_life=20, opponent_life=20):
     attacking_creatures = decision_array[self_creatures, opponent_creatures, self_life, opponent_life]
     count_array[self_creatures, opponent_creatures, self_life, opponent_life] += 1
-    return attacking_creatures
+    count_array = count_array.astype(int)
+    return int(attacking_creatures)
 
 def reward(decision_array, count_array ,winner=None): #currently self = player 2, opponent = player 1
+    count_array = count_array.astype(int)
     boolArr = count_array != 0 
     if winner == True:
         decision_array[boolArr] += count_array[boolArr]
@@ -45,6 +47,7 @@ def reward(decision_array, count_array ,winner=None): #currently self = player 2
         decision_array[boolArr] = decision_array[boolArr] - count_array[boolArr]
         for i in range(n):
             np.clip(decision_array[i,:,:,:], 0, i, out = decision_array[i,:,:,:])
+    decision_array = decision_array.astype(int)
 
 
 # In[ ]:
